SB16 driver + DOOM sound module
- sb16.c: Low-level SB16 driver (DSP reset, version, mixer, DMA,
  auto-init playback, IRQ handler stub)
- sb16.h: SB16 I/O ports, register commands, structs, prototypes
  (fixed dma_buffer_size uint16_t -> uint32_t)
- doom_nyxos_sound.c: DOOM sound/music module (DMX lump parser,
  channel management, Gus_* callbacks)
- doomfeatures.h: #define FEATURE_SOUND 1
- kernel.c: sb16_init() call in boot sequence after speaker_init()
- makefile: Added sb16.o, doom_nyxos_sound.o with DOOM_CFLAGS

+++ kernel/sb16.h
@@ -45,7 +45,7 @@ typedef struct {
     uint8_t  dma16;
     uint8_t  initialized;
     uint32_t sample_rate;
-    uint16_t dma_buffer_size;
+    uint32_t dma_buffer_size;
     uint8_t* dma_buffer;
     uint32_t dma_buffer_phys;
 } sb16_t;
